传感器传值公式变更。

+++ code/src/main/java/cn/com/nex/monitor/webapp/dashboard/bean/RadiationBean.java
@@ -8,6 +8,6 @@ import lombok.EqualsAndHashCode;
 @EqualsAndHashCode(callSuper = false)
 public class RadiationBean extends CommonBean {
     private int radNo;
-    private int radValue;
+    private double radValue;
     private String uploadTime;
 }
